extract router.php from phar, before starting the server
